% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expppel-package.R
\docType{package}
\name{expppel-package}
\alias{expppel}
\alias{expppel-package}
\title{expppel: Prepare cartridge case x3p scans for comparison}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Joe Zemmels \email{jzemmels@iastate.edu}

Authors:
\itemize{
  \item Heike Hofmann \email{hofmann@iastate.edu}
  \item Susan VanderPlas \email{susan.vanderplas@unl.edu}
}

}
\keyword{internal}
